+++ src/index.scss
@@ -1,0 +1,11 @@
+@import './css/normalize.css', './css/skeleton.css';
+
+li.courses-list {
+  list-style: none;
+  border: 1px solid #ccc;
+  padding: 1em;
+
+  h3 {
+    font-size: 2rem;
+  }
+}
