+++ src/models/index.ts
@@ -1,0 +1,8 @@
+export interface CourseModel {
+  id: number;
+  title: string;
+  description: string;
+  authorID: string;
+  date: Date;
+  rating: number;
+}
